==> r-code/01_rawdata.Rmd <==
---
title: "rawdata"
output: pdf_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(readxl)
```

```{r}
## warnings: if plot_id eg. 169.0 -> modified by google docs - open with excel
sites                <- read_excel("data/rawdata/Neofitas_sites.xlsx", sheet = "sites") 
neophytes            <- read_excel("data/rawdata/Neofitas_species.xlsx", sheet = "neophytes")
taxonomy             <- read_excel("data/rawdata/Neofitas_species.xlsx", sheet = "taxonomy")
plantTraitsHierarchy <- read_excel("data/rawdata/TRY_dataset/Neophytes_traitsR.xlsx", sheet = "plant_traits_hierarchy")
plantTraits          <- read_excel("data/rawdata/TRY_dataset/Neophytes_traitsR.xlsx", sheet = "plant_traits")
```

Check traits:
unique(plantTraits[which(!plantTraits$TraitName %in% plantTraitsHierarchy$TraitName),2])
unique(plantTraits[which(!plantTraits$Trait %in% plantTraitsHierarchy$Trait),c(2,3)])
unique(plantTraits[which(!plantTraits$Trait_id %in% plantTraitsHierarchy$Trait_id),c(2,3,4)])


```{r}
save(
  sites,neophytes,taxonomy,
  plantTraits,plantTraitsHierarchy,
  file="data/rawdata.RData"
)
```
