==> r-code/02_workingdata.Rmd <==
---
title: "workingdata"
output: pdf_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
source("r-code/00_functions.R")
load("data/rawdata.RData")
using("magrittr","dplyr","labdsv", "simba","tidyr","vegan","BiodiversityR","hablar","caret","fuzzySim","caret","psych","ggcorrplot","ggcorrplot2","sjPlot") 
# simba: #bb2num
# hablar:retype(), simplyfiing object into the following classes: date, numeric and character
# labdsv: matrify
```



# aspect value
http://ecology.msu.montana.edu/labdsv/R/labs/lab2/lab2.html

```{r}
sites$aspect=(cos((sites$exposure-30)/180*pi)+1)/2
sites$aspect[which(is.na(sites$aspect))]= 0.5 # neither hot or cold
```


Reshape species matrix and convert to percent

```{r}
# https://stackoverflow.com/questions/50691393/transform-community-data-into-wide-format-for-vegan-package
#df = species %>% select(plot_id,species,cover)
#dat.bb = matrify(as.matrix(df))

df = neophytes %>% dplyr::select(site_id,ring,plot_id,plant_id,cover)
dat.bb = df %>% tidyr::spread(key=plant_id,value=cover)

dat.bb= dat.bb %>% dplyr::mutate_at(c(4:ncol(dat.bb)), ~replace(., is.na(.), 0)) # NO factors before!

neophytesPlotPercent=dat.bb
neophytesPlotPercent[,4:ncol(neophytesPlotPercent)]= bb2num(neophytesPlotPercent[,4:ncol(neophytesPlotPercent)], from = c("*", "1", "2", "3", "4", "5"), to = c(0.5, 2.5, 15, 37.5, 62.5, 87.5))
```


# Creation of full dataset: veg
veg dataset incudes also plots with no neophytes
```{r}
# 1 check if all plots of neophytes are present in sites
#neophytes[which(!neophytes$plot_id %in% sites$plot_id),]

# 2 Generate 0 for all plots with no neofites records
sites[which(!sites$plot_id %in% neophytes$plot_id),]

siteNoNeophytes = sites[which(!sites$plot_id %in% neophytesPlotPercent$plot_id),1:3]
NoNeophytes= neophytesPlotPercent[1:nrow(siteNoNeophytes),-c(1:3)]
NoNeophytes[NoNeophytes>0]=0
df= data.frame(cbind(siteNoNeophytes,NoNeophytes))
colnames(df)=colnames(neophytesPlotPercent)
veg=rbind(neophytesPlotPercent,df)
```


# Environment - Dummy variables
```{r}
# scale continous environmental variables
env=sites %>% dplyr::select(-exposure)
env %<>% hablar::retype() 
cols=c("traffic","soil_type","humus","treeAbund","shrubAbund","herbAbund")
env %<>% dplyr::mutate_each_(funs(factor(.)),cols)


dat = env %>% dplyr::select(ring,traffic,plant_comunity,soil_type,humus,treeAbund,herbAbund,shrubAbund) # colSums(is.na(df))
dummies_model <- caret::dummyVars( ~ ., data=dat) # # Creating dummy variables is converting a categorical variable to as many binary variables as here are categories
datDummy = predict(dummies_model, newdata = dat)
colnames(datDummy)[1:4] = sub("\\..","",colnames(datDummy)[1:4])

datBinNum = env %>% dplyr::select(site_id,plot_id,activity,elevation,slope,leak,aspect,soil_coverage)
env<- data.frame(datBinNum,datDummy)

# activity
env$activity=as.character(env$activity)
env$activity[env$activity=="inactive"]=0
env$activity[env$activity=="active"]=1
env$activity=as.numeric(env$activity)

# soil_coverage
env$soil_coverage=as.character(env$soil_coverage)
env$soil_coverage[env$soil_coverage=="low-medium"]=0
env$soil_coverage[env$soil_coverage=="high"]=1
env$soil_coverage=as.numeric(env$soil_coverage)

env4cor = env[,-c(1,2)] # only as numeric needed!
```



```{r}
# correlation
predNamesAbbr =data.frame(names=colnames(env4cor),abbr=c("act","alt","slo","lea","asp","scov","r1","r2", "r3","r4","rref","t0","t1","t2","t3","pcAf","pcF","pcP","pcS","st1","st2","st3","st4","hum0","hum1","hum2","tA0","tA1","tA2","tA3","tA4","hA0","hA1","hA2","hA3","hA4","sA0","sA1","sA2","sA3","sA4"))

envAbbr = env4cor
colnames(envAbbr)=dplyr::coalesce(predNamesAbbr$abbr[match(colnames(envAbbr), predNamesAbbr$names)], colnames(envAbbr))


# ct <- psych::corr.test(env[,-c(1,2)], adjust = "none")
# corr <- ct$r
# p.mat <- ct$p
# ggcorrplot2::ggcorrplot.mixed(corr)
# 
# 
# ggcorrplot::ggcorrplot(corr, hc.order = TRUE, type = "lower",
#    outline.col = "white",
#    ggtheme = ggplot2::theme_gray,p.mat = p.mat,insig = "blank",
#    colors = c("#6D9EC1", "white", "#E46726"))


# # also maybe:
# sjPlot::sjp.corr(envAbbr,decimals=1)


# matrix2.cor=cor(env4cor)
# zzz<-rep(0,ncol(matrix2.cor))
# z=1
# 
# for (i in 1:ncol(matrix2.cor)){
# if(
# (length(which(abs(matrix2.cor[-z,i])>=0.5))==0)==FALSE
# ){ # integer(0)
# zzz[i]<-paste(names(which(abs(matrix2.cor[-z,i])>=0.5)),collapse=",")
# print(paste(names(which(abs(matrix2.cor[-z,i])>=0.5)),collapse=","))
# }
# 
# z=z+1
# }
# predictors.cor<-data.frame(zzz)
# rownames(predictors.cor)<-colnames(matrix2.cor)
# colnames(predictors.cor)<-"r(s) >= 0.5"
# 
# pred.uncor = PredictorSelection(predictors.cor)

pred.uncor=c(
"elevation","slope","aspect","ring2","ring3","ring4",                       
"traffic.2","plant_comunityaltered.forest","soil_type.1",                
"soil_type.2","soil_type.3", "soil_type.4", 
"treeAbund.1","treeAbund.2","treeAbund.3",        
"treeAbund.4","herbAbund.0","herbAbund.1",          
"herbAbund.2","herbAbund.3","shrubAbund.0",               
"shrubAbund.1","shrubAbund.2","shrubAbund.3",              
"activity","ring1","herbAbund.4",            
"traffic.0","shrubAbund.4","humus.0")


## selection
# activity,ring1,traffic.0,shrubAbund.4,humus.0

envAbbr = env4cor[,which(colnames(env4cor) %in% pred.uncor)]
colnames(envAbbr)=dplyr::coalesce(predNamesAbbr$abbr[match(colnames(envAbbr), predNamesAbbr$names)], colnames(envAbbr))

# ct <- psych::corr.test(envAbbr, adjust = "none")
# corr <- ct$r
# p.mat <- ct$p
# ggcorrplot2::ggcorrplot.mixed(corr)
#### Correlation End ###########
```

```{r}
preds = as_tibble(env[,c(1,2,which(colnames(env) %in% pred.uncor))])
cols=c("activity","ring1","ring2","ring3","ring4","traffic.0","traffic.2",
"plant_comunityaltered.forest", "soil_type.1","soil_type.2","soil_type.3","soil_type.4","humus.0",
"treeAbund.1","treeAbund.2","treeAbund.3","treeAbund.4",                 
"herbAbund.0","herbAbund.1","herbAbund.2","herbAbund.3","herbAbund.4",
"shrubAbund.0","shrubAbund.1","shrubAbund.2","shrubAbund.3","shrubAbund.4")
preds %<>% dplyr::mutate_each_(funs(factor(.)),cols)

preds[,-1] = preds[,-1] %>% dplyr::mutate_if(is.numeric, function(x, na.rm = FALSE) as.numeric(scale(x))) # site_id is numeric
preds = as_tibble(preds)
```



# Diversity - inverse Simpson
## Simpson 
Simpson’s Index (D) measures the probability that two individuals randomly selected from a sample will belong to the same species

Simpson’s Reciprocal Index 1 / D. (D2)

The value of this index starts with 1 as the lowest possible figure. This figure would represent a community containing one species. The higher the value, the greater the diversity. The maximum value is the number of species (or other category being used) in the sample. For example if there are five species in the sample, then the maximum value is 5.
https://archetypalecology.wordpress.com/2018/02/20/biodiversity-indices-concepts-and-r-implementations-in-progress/

```{r}
df =veg[,-c(1:3)]
df[df>0] =1
Nneophytes = apply(df,1,sum)

Pneophytes = rep(0,length(Nneophytes))
Pneophytes[which(Nneophytes>0)] = 1

Dneophytes = vegan::diversity(df, index="invsimpson")
Dneophytes[which(is.infinite(Dneophytes))] = NA

dat = data.frame(cbind(veg[,1:3],Pneophytes,Nneophytes,Dneophytes))
vegEnv = as_tibble(dplyr::left_join(dat,env,by=c("site_id","plot_id")))
```



## Presences per site
```{r}
s=neophytesPlotPercent[,4:ncol(neophytesPlotPercent)]
s[s>0]=1 # function sign
plants=as_tibble(cbind(neophytesPlotPercent[,1],s))
vegSite=plants %>% dplyr::group_by(site_id) %>% summarise_all("max")
```




# preparation of datasets for statistical analyses: presence of 6 neophytes (<5%) + species richness (total number)
- presence/absence of neophytes
- total number of neophytes
```{r}
s=veg[,4:ncol(veg)]
s[s>0]=1 # function sign
NtotalNeophytes=apply(s,1,sum) # total number of all neophytes


s=vegan::specnumber(t(veg[,4:ncol(veg)]))
# > s
#Aru.don Cyn.dac Cyp.iri Ele.ind Eup.hir Leo.jap Leu.leu Meg.max Mel.alb Psi.gua Ric.com Son.asp Sor.hal 
#      3      18       9       9      14       1       1      78       2       1      18       2       3

df=round(s/nrow(veg)*100,2)
df
#Aru.don Cyn.dac Cyp.iri Ele.ind Eup.hir Leo.jap Leu.leu Meg.max Mel.alb Psi.gua Ric.com Son.asp Sor.hal 
#   2.10   12.59    6.29    6.29    9.79    0.70    0.70   54.55    1.40    0.70   12.59    1.40    2.10

# df[df>9]
# Cyn.dac Eup.hir Meg.max Ric.com 
#   12.59    9.79   54.55   12.59

selectedNeophytes =veg %>% dplyr::select(site_id,ring,plot_id,names(df[df>10]))


# presences and total richness for selected neophytes
s=selectedNeophytes[,4:ncol(selectedNeophytes)]
s[s>0]=1 # function sign
NfreqNeophytes=apply(s,1,sum) # total number of selected neophytes

plants = as_tibble(cbind(veg[1:3],s,NfreqNeophytes,NtotalNeophytes))
```

# Traits - correlation

## Input for heat tree matrix
```{r}
sites$plot = paste("plot_",sites$plot_id,sep="")
# which plots have neophyte data?
NeophyteSites = sites[-which(!sites$plot_id %in% neophytesPlotPercent$plot_id),] 
NeophyteSites=NeophyteSites[-which(NeophyteSites$traffic==0),] # ignoring reference sites, only 2 records

# compile trait data
Traits = dplyr::inner_join(plantTraits,plantTraitsHierarchy,by=c("Trait_id"="Trait_id"))
Traits %<>% select(plant_id,Trait_id,hierarchy)

# reshape dataset: rows = trait, cols = plots
veg_data = data.frame(neophytesPlotPercent)
rownames(veg_data) = paste("plot_",veg_data$plot_id,sep="")

# remove plots that are not in NeophyteSites
veg_data=veg_data[-which(!rownames(veg_data) %in% NeophyteSites$plot),]

veg_data = veg_data %>% dplyr::select(-site_id,-ring,-plot_id)
veg_dataT = transpose_df(veg_data)
colnames(veg_dataT)[1] = "plant_id"

# combine veg and trait datasets
veg_trait_data = dplyr::inner_join(Traits,veg_dataT,by=c("plant_id"="plant_id"))
```



# Correlation Traits
## Correlation between traits
```{r}
df = veg_trait_data %>% dplyr::select(-plant_id,-hierarchy)
df = data.frame(df %>% group_by(Trait_id)%>% summarise_all(sum))
rownames(df) = df$Trait_id
df = df[,-1]
df = data.frame(t(df))

# correlation

# ct <- psych::corr.test(df, adjust = "none")
# corr <- ct$r
# p.mat <- ct$p
# ggcorrplot2::ggcorrplot.mixed(corr)


# ggcorrplot::ggcorrplot(corr, hc.order = TRUE, type = "lower",
#    outline.col = "white",
#    ggtheme = ggplot2::theme_gray,p.mat = p.mat,insig = "blank",
#    colors = c("#6D9EC1", "white", "#E46726"))
# 
# 
# # # also maybe:
# # sjPlot::sjp.corr(envAbbr,decimals=1)



#### Correlation End ###########
```


## Correlation for traits for non-singleton neophytes
```{r}
# SELECT non-singleton neophytes
names=c("Cyn.dac","Cyp.iri","Ele.ind", "Eup.hir","Meg.max","Mel.alb","Ric.com","Son.asp","Sor.hal")
spe4c = neophytesPlotPercent %>% dplyr::select(site_id,plot_id,all_of(names)) 
Traits4c = plantTraits[which(plantTraits$plant_id %in% colnames(spe4c)),] # select neophytes
Traits4c %<>% dplyr::select(plant_id,TraitName,Trait_id)
traitsNoVariance = table(Traits4c$plant_id,Traits4c$Trait_id)
selectedTraits = names(which(apply(traitsNoVariance,2,sum)!=0 & apply(traitsNoVariance,2,sum)!=length(names)))

df = veg_trait_data[which(veg_trait_data$Trait_id %in% selectedTraits),]
df = df %>% dplyr::filter(plant_id %in% names)
df = df %>% dplyr::select(-plant_id,-hierarchy)
df = data.frame(df %>% group_by(Trait_id)%>% summarise_all(sum))
rownames(df) = df$Trait_id
df = df[,-1]
traitsPlot = data.frame(t(df))

# correlation
# ct <- psych::corr.test(traitsPlot, adjust = "none")
# corr <- ct$r
# p.mat <- ct$p
# ggcorrplot2::ggcorrplot.mixed(corr)
# 
# matrix2.cor=cor(traitsPlot)
# zzz<-rep(0,ncol(matrix2.cor))
# z=1
# 
# for (i in 1:ncol(matrix2.cor)){
# if(
# (length(which(abs(matrix2.cor[-z,i])>=0.5))==0)==FALSE
# ){ # integer(0)
# zzz[i]<-paste(names(which(abs(matrix2.cor[-z,i])>=0.5)),collapse=",")
# print(paste(names(which(abs(matrix2.cor[-z,i])>=0.5)),collapse=","))
# }
# 
# z=z+1
# }
# predictors.cor<-data.frame(zzz)
# rownames(predictors.cor)<-colnames(matrix2.cor)
# colnames(predictors.cor)<-"r(s) >= 0.5"
# 
# traits.uncor = PredictorSelection(predictors.cor)
# # selection: TolShadeH,HeightMaxH,Ny,DispS,TolFireL

# traits.uncor
#[1] "TolShadeH"  "HeightMaxH" "Ny"         "DispS"      "ReprodVeg"  "TolFireL" 

traitSelection4c.NS = as.data.frame.matrix(traitsNoVariance[,colnames(traitsNoVariance) %in% traits.uncor])
```


## Correlation for traits for 3 seected neophytes (>10% occ)
```{r}
# SELECT 3 neophytes
names=c("Cyn.dac","Meg.max","Ric.com")
spe4c = neophytesPlotPercent %>% dplyr::select(site_id,plot_id,all_of(names)) 
Traits4c = plantTraits[which(plantTraits$plant_id %in% colnames(spe4c)),] # select neophytes
Traits4c %<>% dplyr::select(plant_id,TraitName,Trait_id)
traitsNoVariance = table(Traits4c$plant_id,Traits4c$Trait_id)
selectedTraits = names(which(apply(traitsNoVariance,2,sum)!=0 & apply(traitsNoVariance,2,sum)!=length(names)))

df = veg_trait_data[which(veg_trait_data$Trait_id %in% selectedTraits),]
df = df %>% dplyr::filter(plant_id %in% names)
df = df %>% dplyr::select(-plant_id,-hierarchy)
df = data.frame(df %>% group_by(Trait_id)%>% summarise_all(sum))
rownames(df) = df$Trait_id
df = df[,-1]
traitsPlot = data.frame(t(df))

# correlation
# ct <- psych::corr.test(traitsPlot, adjust = "none")
# corr <- ct$r
# p.mat <- ct$p
# ggcorrplot2::ggcorrplot.mixed(corr)
# 
matrix2.cor=cor(traitsPlot)
zzz<-rep(0,ncol(matrix2.cor))
z=1

for (i in 1:ncol(matrix2.cor)){
if(
(length(which(abs(matrix2.cor[-z,i])>=0.5))==0)==FALSE
){ # integer(0)
zzz[i]<-paste(names(which(abs(matrix2.cor[-z,i])>=0.5)),collapse=",")
print(paste(names(which(abs(matrix2.cor[-z,i])>=0.5)),collapse=","))
}

z=z+1
}
predictors.cor<-data.frame(zzz)
rownames(predictors.cor)<-colnames(matrix2.cor)
colnames(predictors.cor)<-"r(s) >= 0.5"

traits.uncor = PredictorSelection(predictors.cor)
# # traits.uncor
# # [1] "TolShadeH"  "HeightMaxH" "ReprodVeg"

traitSelection4c.sN = as.data.frame.matrix(traitsNoVariance[,colnames(traitsNoVariance) %in% traits.uncor])
```




# save
```{r}
save(
  neophytes,neophytesPlotPercent,
  sites,taxonomy,
  plantTraits,plantTraitsHierarchy,
  veg,env,vegEnv,vegSite,
  plants,preds,pred.uncor,
  traitSelection4c.sN,traitSelection4c.NS,
  veg_trait_data,
  file="data/workingdata.RData"
)
```
