==> r-code/03_descriptive_analyses.Rmd <==
---
title: "workingdata"
output: pdf_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
source("r-code/00_functions.R") # data_summary
load("data/workingdata.RData")
source("r-code/00_functions.R")
using("ggplot2","ellipse","magrittr","tidyverse","vegan","labdsv","broom","gridExtra","FSA","rstatix")
# ellipse: #library(GGally) problem with NA's
# magrittr: # for %<>%

mytheme <-  theme(axis.title = element_text(size =18),
        axis.text = element_text(size =14),
        strip.text= element_text(size = 18),
        legend.position = "none",
        plot.title = element_text(hjust = 0.5))
```


# 1 How many sites were sampled with how many plots and how many plants were found

apply(plantOccSite[,-c(1)],1,sum)
is the same as following command from biodiversityR package:
diversityresult(plantOccSite[,-c(1)], index = 'richness', method = 'each site')

```{r}
# nr. plants per site
plantsCounts=as_tibble(cbind(vegSite[,1],N_neophytes=apply(vegSite[,-c(1)],1,sum)))

# # nr. plants per ring
# s=speciesPercent[,4:17]
# s[s>0]=1
# plants=as_tibble(cbind(speciesPercent[,1:2],s))
# plantsGrp=plants %>% dplyr::group_by(site_id,ring) %>% summarise_all("max")
# plantsCounts=as_tibble(cbind(plantsGrp[,1:2],counts=apply(plantsGrp[,-c(1:2)],1,sum)))


x=table(sites$site_id,sites$ring);x
df= data.frame(cbind(site_id=rownames(x),x,total=apply(x,1,sum),N_neophytes = plantsCounts$N_neophytes))
colnames(df)[2:5]=1:4;
df[] <- lapply(df, as.numeric)
df
total=apply(df,2,sum)
total[1]=nrow(df)
rbind(df,total=total)
hist(df$N_neophytes,main="Histogram of number of neophytes sample per site",xlab="Number of neophytes sample per site")
```


```{r}
Nocc = apply(vegSite[,-1],2,sum)

ACT = sites %>% dplyr::group_by(site_id,activity) %>% summarize(n()) %>% dplyr::select(site_id,activity) 
df = dplyr::left_join(vegSite,ACT,by="site_id")
df= df[,c(1,15,2:14)] # rearrange columns

Table1 = data.frame(matrix(0,nrow=13,ncol=3))
rownames(Table1) = colnames(df[,3:ncol(df)])
colnames(Table1) = c("TotalOcc","actOcc","inactOcc")
Table1[,1] = apply(df[,-c(1,2)],2,sum)

dfActGr = data.frame(df %>% select(-site_id) %>% dplyr::group_by(activity) %>% summarise(across(everything(), sum)))
Table1[,2] = as.numeric(dfActGr[1,-1])
Table1[,3] = as.numeric(dfActGr[2,-1])
Table1
```








# 2 Test for differences between ring in number of neophytes and inverse Simpson
```{r}
#1. number of neophytes
# tables needed: sites, vegEnv
df = vegEnv %>% dplyr::select(ring,Nneophytes)
df$ring=as.factor(df$ring)
df%<>% arrange(ring) # sort for ring and activity


########### Kruskal-Wallis #############
# 1) Differences between rings
kruskal.test(Nneophytes ~ ring, data = df)
# Kruskal-Wallis rank sum test
# data:  Nneophytes by ring
# Kruskal-Wallis chi-squared = 58.715, df = 4, p-value = 5.401e-12

#perform Dunn's Test with Bonferroni correction for p-values
dunn_test(Nneophytes ~ ring,data=df,p.adjust.method="bonferroni")
# # A tibble: 10 x 9
#    .y.        group1 group2    n1    n2 statistic             p        p.adj p.adj.signif
#  * <chr>      <chr>  <chr>  <int> <int>     <dbl>         <dbl>        <dbl> <chr>       
#  1 Nneophytes 1      2         18    40     -2.55 0.0109        0.109        ns          
#  2 Nneophytes 1      3         18    34     -4.32 0.0000153     0.000153     ***         
#  3 Nneophytes 1      4         18    35     -6.06 0.00000000136 0.0000000136 ****        
#  4 Nneophytes 1      ref       18    16     -6.10 0.00000000108 0.0000000108 ****        
#  5 Nneophytes 2      3         40    34     -2.31 0.0211        0.211        ns          
#  6 Nneophytes 2      4         40    35     -4.47 0.00000769    0.0000769    ****        
#  7 Nneophytes 2      ref       40    16     -4.64 0.00000348    0.0000348    ****        
#  8 Nneophytes 3      4         34    35     -2.07 0.0388        0.388        ns          
#  9 Nneophytes 3      ref       34    16     -2.75 0.00591       0.0591       ns          
# 10 Nneophytes 4      ref       35    16     -1.12 0.264         1            ns 

#2. inverse Simpson
# tables needed: sites, vegEnv
df = vegEnv %>% dplyr::select(ring,Dneophytes)
df%<>%filter(Dneophytes!=0) # library(magrittr) for %<>%
df$ring=as.factor(df$ring)
df%<>% arrange(ring) # sort for ring and activity

########### Kruskal-Wallis #############
# 1) Differences between rings
df=df %>% dplyr::filter(ring!='ref')
kruskal.test(Dneophytes ~ ring, data = df)
# 	Kruskal-Wallis rank sum test
# data:  Dneophytes by ring
# Kruskal-Wallis chi-squared = 20.166, df = 3, p-value = 0.0001568

#perform Dunn's Test with Bonferroni correction for p-values
dunn_test(Dneophytes ~ ring,data=df,p.adjust.method="bonferroni")
# # A tibble: 6 x 9
#   .y.        group1 group2    n1    n2 statistic         p     p.adj p.adj.signif
# * <chr>      <chr>  <chr>  <int> <int>     <dbl>     <dbl>     <dbl> <chr>       
# 1 Dneophytes 1      2         18    33     -2.38 0.0173    0.104     ns          
# 2 Dneophytes 1      3         18    20     -3.02 0.00254   0.0152    *           
# 3 Dneophytes 1      4         18    13     -4.33 0.0000148 0.0000887 ****        
# 4 Dneophytes 2      3         33    20     -1.00 0.317     1         ns          
# 5 Dneophytes 2      4         33    13     -2.69 0.00724   0.0434    *           
# 6 Dneophytes 3      4         20    13     -1.67 0.0943    0.566     ns
```

# 3 Barplots - Number of neophytes,inverse Simpson and cover per plot depending on site activity
```{r}
# tables needed: vegEnv

## 1. Plot: species richness (= Nneophytes)
df = vegEnv %>% dplyr::select(-site_id,-plot_id)
df%<>%filter(!ring =='ref') # library(magrittr) for %<>%
df$ring=as.factor(df$ring)
df$activity=as.factor(df$activity)
df$value=df$Nneophytes
df%<>% arrange(ring,activity) # sort for ring and activity

########### 2-group Mann-Whitney U Test #############
#wilcox.test(Dneophytes ~ activity, data = df)
# 	Wilcoxon rank sum test with continuity correction
# 
# data:  Dneophytes by activity
# W = 1093.5, p-value = 0.02773
# alternative hypothesis: true location shift is not equal to 0

#df %>% group_by(ring) %>%  do(broom::tidy(kruskal.test(x = .$Dneophytes, g = .$activity))) 
# dat = df %>% dplyr::filter(ring=='1')
# wilcox.test(Dneophytes ~ activity, data = dat) # W = 64, p-value = 0.03816
# dat = df %>% dplyr::filter(ring=='2')
# wilcox.test(Dneophytes ~ activity, data = dat) # W = 184, p-value = 0.05986
# dat = df %>% dplyr::filter(ring=='3')
# wilcox.test(Dneophytes ~ activity, data = dat) # W = 47.5, p-value = 0.887
# dat = df %>% dplyr::filter(ring=='4')
# wilcox.test(Dneophytes ~ activity, data = dat) 

#df %>% group_by(ring,activity) %>% get_summary_stats(Nneophytes, type = "mean_sd")
####################################################



dfPlot <- data_summary(df, varname=c("value"),
                    groupnames=c("ring","activity"))
dfPlot$yminSD<-dfPlot$value-dfPlot$sd
dfPlot$ymaxSD<-dfPlot$value+dfPlot$sd
dfPlot[which(dfPlot$yminSD<0),"yminSD"]<-0

levels(df$activity)=c("inactive","active")
dfPlot$nObs =  paste(levels(df$activity),"\n(N=",as.vector(table(df$activity,df$ring)),")",sep="")


#dfPlot$sig<-  rep("a",nrow(dfPlot))
dfPlot$sig<-c("a","b","a","a","a","a","","")

g1 =  ggplot(dfPlot,aes(x=nObs,y=value,fill=activity)) + geom_bar(stat="identity")+
  geom_errorbar(aes(ymin=yminSD, ymax=ymaxSD), width=.2) +
  #scale_x_discrete(labels=nObs) +
  labs(y = "SR", x = "") + # rename y-axis
    scale_fill_manual(values=rep(c("grey41", "grey79"),5))+
    facet_grid(~ring, scales="free_x")+
  scale_y_continuous(breaks=c(0,1,2,3,4,5)) +
  theme_bw() + mytheme +
  geom_text(mapping = aes( y =6, label = sig),size = 6,  fontface=1)







###############################################

## 2. Plot: inverse Simpson
df = vegEnv %>% dplyr::select(-site_id,-plot_id)
df%<>%filter(Dneophytes!=0 & !ring =='ref') # library(magrittr) for %<>%
df$ring=as.factor(df$ring)
df$activity=as.factor(df$activity)
df$value=df$Dneophytes
df%<>% arrange(ring,activity) # sort for ring and activity


########### 2-group Mann-Whitney U Test #############
#wilcox.test(Dneophytes ~ activity, data = df)
# 	Wilcoxon rank sum test with continuity correction
# 
# data:  Dneophytes by activity
# W = 1093.5, p-value = 0.02773
# alternative hypothesis: true location shift is not equal to 0

#df %>% group_by(ring) %>%  do(broom::tidy(kruskal.test(x = .$Dneophytes, g = .$activity))) 
# dat = df %>% dplyr::filter(ring=='1')
# wilcox.test(Dneophytes ~ activity, data = dat) # W = 64, p-value = 0.03816
# dat = df %>% dplyr::filter(ring=='2')
# wilcox.test(Dneophytes ~ activity, data = dat) # W = 184, p-value = 0.05986
# dat = df %>% dplyr::filter(ring=='3')
# wilcox.test(Dneophytes ~ activity, data = dat) # W = 47.5, p-value = 0.887
# 
# df %>% group_by(ring,activity) %>% get_summary_stats(Dneophytes, type = "mean_sd")
####################################################


# function to let barplot start a 1, since invSimpson starts from 1 (= lowest diversity)
require(scales)
my_trans <- function(from=0) 
{
  trans <- function(x) x-from
  inv <- function(x) x+from
  trans_new("myscale", trans, inv, 
            domain = c(from, Inf))
}

dfPlot <- data_summary(df, varname=c("value"),
                    groupnames=c("ring","activity"))
dfPlot$yminSD<-dfPlot$value-dfPlot$sd
dfPlot$ymaxSD<-dfPlot$value+dfPlot$sd
dfPlot[which(dfPlot$yminSD<1),"yminSD"]<-1

levels(df$activity)=c("inactive","active")
dfPlot$nObs =  paste(levels(df$activity),"\n(N=",as.vector(table(df$activity,df$ring)),")",sep="")


#dfPlot$sig<-  rep("a",nrow(dfPlot))
dfPlot$sig<-c("a","b","a","a","a","a","","")

g2 =  ggplot(dfPlot,aes(x=nObs,y=value,fill=activity)) + geom_bar(stat="identity")+
  geom_errorbar(aes(ymin=yminSD, ymax=ymaxSD), width=.2) +
  #scale_x_discrete(labels=nObs) +
  labs(y = "1/Sim", x = "") + # rename y-axis
    scale_fill_manual(values=rep(c("grey41", "grey79"),5))+
    facet_grid(~ring, scales="free_x")+
    scale_y_continuous(trans = my_trans( from=1),breaks=c(1,2,3,4,5)) +
  theme_bw() + mytheme +
  geom_text(mapping = aes( y =6, label = sig),size = 6,  fontface=1)







grid.arrange(g1, g2, nrow = 2)
```





















# 4 Number of neophytes per plot depending on site activity
```{r}
# tables needed: sites, veg
x=table(sites$ring,sites$activity);x # N plots per ring and activity

vegRingActivity = merge(x = veg, y = sites[,c("site_id","ring","plot_id","activity")], by = c("site_id","ring","plot_id"), all.x = TRUE)

df= vegRingActivity %>% dplyr::select(-site_id,-plot_id)

df$ring=as.factor(df$ring)
df$activity=as.factor(df$activity)

vegCoverMeanRingActivity = df %>% dplyr::group_by(ring,activity)%>% dplyr::summarise_if(is.numeric, mean)
vegOccMeanRingActivity = vegCoverMeanRingActivity %>% select(-ring,activity) %>% mutate_if(is.numeric, ~1 * (. > 0)) # all values >0 to zero


# barplot: N species per ring and activity
df = vegOccMeanRingActivity[,1:2]
df$N_neophytes = vegan::specnumber(vegOccMeanRingActivity[,-c(1:2)])

df$nObs <- paste(levels(df$activity),"\n(N=",as.vector(table(vegRingActivity$activity,vegRingActivity$ring)),")",sep="")


ggplot(df,aes(x=nObs,y=N_neophytes,fill=activity)) + geom_bar(stat="identity") + 
    #ggtitle(paste("Number of neophyte presences per ring and petrol activity"))+
    labs(y = "Number of neophyte presences", x = "") + # rename y-axis
    scale_fill_manual(values=rep(c("grey41", "grey79"),5))+
    facet_grid(~ring, scales="free_x")+
    scale_x_discrete(breaks=df$nObs,labels=df$nObs) +
   theme_bw() + mytheme


```






# Descriptive figure - Number of neophytes presences per zone
- attention: this plot maybe wrong due to unequal number of plots
```{r}
df = neophytesPlotPercent
dfGather = df %>% gather(key = "species", value = "cover", 4:ncol(df))

dfPlot<- dfGather %>% dplyr::filter(cover!=0) %>% dplyr::count(ring,species)

ggplot(dfPlot,aes(x=species,y=n))+ geom_bar(stat="identity")+
    facet_wrap(~ring, ncol=5)+
  labs(x= "", y = "Number of presences") + 
  theme(axis.text.x = element_text(angle = 90, 
                                   hjust = 0, vjust=0.4,face = "italic"))
```






# Descriptive figure - Number of neophytes presences per zone
- attention: this plot maybe wrong due to unequal number of plots
```{r}
df1 = neophytes %>% dplyr::select(site_id,ring,plot_id,plant_id,cover)
df2= sites%>% dplyr::select("site_id","plot_id","activity")
dfC = dplyr::left_join(df1,df2,by=c("site_id","plot_id"))

#table(sites$ring,sites$activity)
  #     active inactive
  # 1        9        9
  # 2       20       20
  # 3       17       17
  # 4       17       18
  # ref      6       10

colnames(dfC)[c(4,5)]=c("plant_id","cover")
dfPlot = dfC%>% group_by(plant_id,ring,cover,activity)%>% summarise(n = n())
dfPlot$freq = dfPlot$n
dfPlot$freq[dfPlot$ring=="1" ] = dfPlot$freq[dfPlot$ring=="1" ]/9
dfPlot$freq[dfPlot$ring=="2" ] = dfPlot$freq[dfPlot$ring=="2" ]/20
dfPlot$freq[dfPlot$ring=="3" ] = dfPlot$freq[dfPlot$ring=="3" ]/17
dfPlot$freq[dfPlot$ring=="4" & dfPlot$activity=="active"] = dfPlot$freq[dfPlot$ring=="4" & dfPlot$activity=="active"]/17
dfPlot$freq[dfPlot$ring=="4" & dfPlot$activity=="inactive"] = dfPlot$freq[dfPlot$ring=="4" & dfPlot$activity=="inactive"]/18
dfPlot$freq[dfPlot$ring=="ref" & dfPlot$activity=="active"] = dfPlot$freq[dfPlot$ring=="ref" & dfPlot$activity=="active"]/6
dfPlot$freq[dfPlot$ring=="ref" & dfPlot$activity=="inactive"] = dfPlot$freq[dfPlot$ring=="ref" & dfPlot$activity=="inactive"]/10

dfPlot %>% filter(plant_id %in% c("Cyn.dac","Meg.max")) %>%
dfPlot %>% 
ggplot( aes(x=cover, y=freq, fill=activity)) +
geom_bar(stat='identity', position=position_dodge2(preserve = "single"))+
    facet_grid(plant_id~ring)+
  labs(x= "", y = "Relative frequencies of cover") + 
  theme(axis.text.x = element_text(angle = 90, 
                                   hjust = 0, vjust=0.4,face = "italic"),legend.position = "none")
```



annotation: for boxplots
nObs <- paste(levels(df$activity),"\n(N=",as.vector(table(df$activity,df$ring)),")",sep="")
df$nObs = rep(nObs,as.vector(table(df$activity,df$ring)))
