==> r-code/05_1_statistical_analyses_preSelection.Rmd <==
---
title: "workingdata"
output: pdf_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
source("r-code/00_functions.R")
load("data/workingdata.RData")
using("plyr","ggplot2","tidyverse","purrr","dplyr","broom","knitr","skimr","MuMIn","ROCR","pscl","pROC","PerformanceAnalytics","randomForest")

rm(.Random.seed, envir=globalenv())
Nplants=3
```



# PreSelection of environmental variables
## Run all combinations of variables for all selected neophytes
```{r}
plantsDivEnv = as_tibble(dplyr::left_join(plants,preds,by=c("site_id","plot_id")))
Nplants=3
plantsEnv = list()
for (i in 1:Nplants){
  plantsEnv[[i]] = plantsDivEnv[,c(3+i,9:ncol(plantsDivEnv))]
  colnames(plantsEnv[[i]])[1] ="y"
  plantsEnv[[i]]$y=as.factor(plantsEnv[[i]]$y)}

predNames=names(plantsDivEnv)[-c(1:8)]
var.predict<-paste(predNames,collapse="+")
rf.form <- as.formula(paste("y",var.predict, sep = " ~ "))

resultsRF = data.frame(matrix(0,nrow=length(predNames),ncol=4))
colnames(resultsRF) = c("predictor",colnames(plantsDivEnv)[4:6])
resultsRF[,1]=predNames

for (i in 1:Nplants){
df=plantsEnv[[i]]
#To make our formula for RF easier to manipulate

#print(rf.form) #This is our current itteration of the formula we're using in RF
set.seed(2021)
plants.rf<-randomForest(rf.form,data=df,importance=TRUE)
#varImpPlot(plants.rf) #Examine our Variable importance plot
resultsRF[,i+1]=plants.rf$importance[,3]
}
resultsRF$meanVImp = apply(resultsRF[,2:4],1,mean)

resultsRF %>%
  mutate(predictor = fct_reorder(predictor,meanVImp)) %>%
ggplot(aes(x=meanVImp,y=predictor))+geom_point()
```

Topology: slope
Anthropogenic disturbance: ring1
vegetation: herbAbund.4
Soil: humus.0
