==> r-code/05_2_statistical_analyses_BGM.Rmd <==
---
title: "Boosted regression trees"
output: pdf_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
source("r-code/00_functions.R")
load("data/workingdata.RData")
using("ggplot2","tidyverse","purrr","dplyr","broom","knitr","dismo","skimr","caret","poibin")

rm(.Random.seed, envir=globalenv())
set.seed(2020)

mytheme <-  theme(axis.title = element_text(size =18),
        axis.text = element_text(size =14),
        strip.text= element_text(size = 18),
        legend.position = "none",
        plot.title = element_text(hjust = 0.5))
```

https://rspatial.org/raster/sdm/9_sdm_brt.html
https://statistik-dresden.de/archives/14967
http://uc-r.github.io/gbm_regression

Input data:
env = environmental variables
plants = selected neophytes + species richness


```{r}
Preds = preds %>% dplyr::select(site_id,plot_id,ring1,slope,herbAbund.4,humus.0)
GbmPlantsPreds = as_tibble(dplyr::left_join(plants,Preds,by=c("site_id","plot_id")))
```






# Tuning of parameters
bag.fraction: the fraction of the training set observations randomly selected to propose the next tree in the expansion
shrinkage/learning.rate: Die „Lerngeschwindigkeit“ – betrifft die Gewichtung schlecht vorhergesagter Fälle für den nachfolgenden Baum. 
interaction.depth/tree.complextity: Die „Tiefe“ der Bäume, d. h. die Anzahl der Ebenen. (Der Beispiel-Baum ganz oben enthält zwei Verzweigungsebenen.) 


```{r}
# create hyperparameter grid
hyper_grid <- expand.grid(
  learning.rate = c(0.005,0.001,0.01),
  bag.fraction = c(.5,.65,.75), 
  tree.complexity  = c(1,3,5),
  optimal_trees = 0,                # a place to dump results
  min_RMSE = 0,                     # a place to dump results
  cv.deviance.mean = 0,             # a place to dump results
  cv.deviance.se = 0,               # a place to dump results
  cv.correlation.mean = 0,          # a place to dump results
  cv.correlation.se = 0            # a place to dump results
)
```



# Creation of gbms based on hypergrid for all seleceted neophytes and species richness
Species richness based on 3 neophytes + 2 SR


```{r}
df = data.frame(GbmPlantsPreds[,-c(1,2,3)]) # without site_id

gbm.tune.result=list()
bgm.family ="bernoulli" # for individual neophytes

for(s in 1:5) {
 if(s %in% c(4,5)) {bgm.family = "poisson"}
 

for(i in 1:nrow(hyper_grid)) {
  
  # reproducibility
  set.seed(2020)
  
    # train model
  gbm.tune <- gbm.step(
    data=df, gbm.x = 6:ncol(df), gbm.y = s,
    family = bgm.family,
    learning.rate = hyper_grid$learning.rate[i],
    bag.fraction = hyper_grid$bag.fraction[i],
    tree.complexity = hyper_grid$tree.complexity[i],
    n.folds=10,
    verbose=T
  )
  
  hyper_grid$optimal_trees[i]       <- ifelse(is.null(gbm.tune$n.trees),NA , gbm.tune$n.trees) # which.min(gbm.tune$train.error) # =
  hyper_grid$min_RMSE[i]            <- ifelse(is.null(sqrt(min(gbm.tune$train.error))),NA ,sqrt(min(gbm.tune$train.error)))
  hyper_grid$cv.deviance.mean[i]    <- ifelse(is.null(gbm.tune$cv.statistics$deviance.mean),NA ,gbm.tune$cv.statistics$deviance.mean)
  hyper_grid$cv.deviance.se[i]      <- ifelse(is.null(gbm.tune$cv.statistics$deviance.se),NA ,gbm.tune$cv.statistics$deviance.se)
  hyper_grid$cv.correlation.mean[i] <- ifelse(is.null(gbm.tune$cv.statistics$correlation.mean),NA ,gbm.tune$cv.statistics$correlation.mean)
  hyper_grid$cv.correlation.se[i]   <- ifelse(is.null(gbm.tune$cv.statistics$correlation.se),NA ,gbm.tune$cv.statistics$correlation.se)
  hyper_grid$AUCtrain.mean[i]       <- ifelse(is.null(gbm.tune$cv.statistics$discrimination.mean),NA , gbm.tune$cv.statistics$discrimination.mean)
  hyper_grid$AUCtrain.se[i]         <- ifelse(is.null(gbm.tune$cv.statistics$discrimination.se),NA , gbm.tune$cv.statistics$discrimination.se)

}
print(hyper_grid)

#gbm.result = hyper_grid %>% arrange(-desc(min_RMSE)) %>% dplyr::filter(optimal_trees > 999)
gbm.result= hyper_grid %>% arrange(-desc(min_RMSE)) 
gbm.tune.result[[s]] = data.frame(gbm.result)
}


gbm.tune.result.df=as_tibble(data.frame(cbind(y=rep(1:5,each=nrow(gbm.tune.result[[1]])),plyr::rbind.fill(gbm.tune.result)))) 
gbm.tune.result.tree1000 = gbm.tune.result.df %>% dplyr::filter(optimal_trees > 999)
print(gbm.tune.result.tree1000,n=100)
```




```{r}
gbm.tune.result.1 = data.frame(matrix(0,ncol=13,nrow=5))
Names= c("y","learning.rate","bag.fraction","tree.complexity","optimal_trees", "min_RMSE","explainedDeviance","cv.deviance.mean", 
"cv.deviance.se","cv.correlation.mean","cv.correlation.se","AUCtrain.mean","AUCtrain.se")
names(gbm.tune.result.1) = Names

fittedValues = as_tibble(data.frame(matrix(0,ncol=3,nrow=143)))

gbm.fits = list()


# plant1
s=1
df=data.frame(GbmPlantsPreds[,c(s+3,9:ncol(GbmPlantsPreds))]) # without site_id

rm(.Random.seed, envir=globalenv())
set.seed(2020)     # reproducibility
    # train model
  gbm.tune <- gbm.step(
    data=df, gbm.x = 2:ncol(df), gbm.y = 1,
    family = "bernoulli",
    learning.rate = 0.0001,
    bag.fraction = 1,
    tree.complexity = 5,
    n.folds=10,
    verbose=F,
    n.trees=1000,
    step.size=1,
    max.trees=50000,
    keep.fold.models = TRUE, keep.fold.vector = TRUE, keep.fold.fit = TRUE
)
   
learning.rate    = gbm.tune$shrinkage
bag.fraction     = gbm.tune$bag.fraction
tree.complexity  = gbm.tune$interaction.depth
optimal_trees           <- ifelse(is.null(gbm.tune$n.trees),NA , gbm.tune$n.trees) # which.min(gbm.tune$train.error) # =
min_RMSE                <- ifelse(is.null(sqrt(min(gbm.tune$train.error))),NA ,sqrt(min(gbm.tune$train.error)))
cv.deviance.mean        <- ifelse(is.null(gbm.tune$cv.statistics$deviance.mean),NA ,gbm.tune$cv.statistics$deviance.mean)
explainedDeviance       <- (gbm.tune$self.statistics$mean.null-gbm.tune$cv.statistics$deviance.mean)/gbm.tune$self.statistics$mean.null
cv.deviance.se          <- ifelse(is.null(gbm.tune$cv.statistics$deviance.se),NA ,gbm.tune$cv.statistics$deviance.se)
cv.correlation.mean     <- ifelse(is.null(gbm.tune$cv.statistics$correlation.mean),NA ,gbm.tune$cv.statistics$correlation.mean)
cv.correlation.se       <- ifelse(is.null(gbm.tune$cv.statistics$correlation.se),NA ,gbm.tune$cv.statistics$correlation.se)
AUCtrain.mean           <- ifelse(is.null(gbm.tune$cv.statistics$discrimination.mean),NA , gbm.tune$cv.statistics$discrimination.mean)
AUCtrain.se             <- ifelse(is.null(gbm.tune$cv.statistics$discrimination.se),NA , gbm.tune$cv.statistics$discrimination.se)

  
 # 5 wwarings glm.fit: fitted probabilities numerically 0 or 1 occurred
gbm.tune.result.1[s,] = c(s,learning.rate,bag.fraction,tree.complexity,optimal_trees,min_RMSE,explainedDeviance,cv.deviance.mean,cv.deviance.se,
                          cv.correlation.mean,cv.correlation.se,AUCtrain.mean,AUCtrain.se)
 
fittedValues[,s] = gbm.tune$fitted
gbm.fits[[s]] = gbm.tune  

# save fitted values of 10 folds - 1 neophyte
for (i in 1:10){
  fold<- paste("fold", sprintf("%02d", i), sep = "_")
  assign(fold,data.frame(cbind(site_id=as.numeric(names(gbm.tune$fold.models[[i]]$data$y)),
    y=gbm.tune$fold.models[[i]]$data$y,fitted=plogis(gbm.tune$fold.models[[i]]$fit))))
}
neophyte1.obsfitted.folds = list(fold_01,fold_02,fold_03,fold_04,fold_05,fold_06,fold_07,fold_08,fold_09,fold_10)



# plant2
s=2
df=data.frame(GbmPlantsPreds[,c(s+3,9:ncol(GbmPlantsPreds))]) # without site_id
tuneParameters=gbm.tune.result.tree1000[gbm.tune.result.tree1000$y==s,]
tuneParameter=tuneParameters[1,] # test individual settings!

rm(.Random.seed, envir=globalenv())
set.seed(2020)     # reproducibility
    # train model
  gbm.tune <- gbm.step(
    data=df, gbm.x = 2:ncol(df), gbm.y = 1,
    family = "bernoulli",
    learning.rate = tuneParameter$learning.rate,
    bag.fraction = tuneParameter$bag.fraction,
    tree.complexity = tuneParameter$tree.complexity,
    n.folds=10,
    verbose=F,
    keep.fold.models = TRUE, keep.fold.vector = TRUE, keep.fold.fit = TRUE
)

# ok!
explainedDeviance=(gbm.tune$self.statistics$mean.null-gbm.tune$cv.statistics$deviance.mean)/gbm.tune$self.statistics$mean.null
gbm.tune.result.1[s,] = c(tuneParameter[1:6],explainedDeviance,tuneParameter[7:12])
fittedValues[,s] = gbm.tune$fitted
gbm.fits[[s]] = gbm.tune  

# save fitted values of 10 folds - 2 neophyte
for (i in 1:10){
  fold<- paste("fold", sprintf("%02d", i), sep = "_")
  assign(fold,data.frame(cbind(site_id=as.numeric(names(gbm.tune$fold.models[[i]]$data$y)),
    y=gbm.tune$fold.models[[i]]$data$y,fitted=plogis(gbm.tune$fold.models[[i]]$fit))))
}
neophyte2.obsfitted.folds = list(fold_01,fold_02,fold_03,fold_04,fold_05,fold_06,fold_07,fold_08,fold_09,fold_10)


# plant3
s=3
df=data.frame(GbmPlantsPreds[,c(s+3,9:ncol(GbmPlantsPreds))]) # without site_id
tuneParameters=gbm.tune.result.tree1000[gbm.tune.result.tree1000$y==s,]
tuneParameter=tuneParameters[1,] # test individual settings!

rm(.Random.seed, envir=globalenv())
set.seed(2020)     # reproducibility
    # train model
  gbm.tune <- gbm.step(
    data=df, gbm.x = 2:ncol(df), gbm.y = 1,
    family = "bernoulli",
    learning.rate = tuneParameter$learning.rate,
    bag.fraction = tuneParameter$bag.fraction,
    tree.complexity = tuneParameter$tree.complexity,
    n.folds=10,
    verbose=F,
    keep.fold.models = TRUE, keep.fold.vector = TRUE, keep.fold.fit = TRUE
)
# ok!
explainedDeviance=(gbm.tune$self.statistics$mean.null-gbm.tune$cv.statistics$deviance.mean)/gbm.tune$self.statistics$mean.null
gbm.tune.result.1[s,] = c(tuneParameter[1:6],explainedDeviance,tuneParameter[7:12])
fittedValues[,s] = gbm.tune$fitted
gbm.fits[[s]] = gbm.tune 

# save fitted values of 10 folds - 3 neophyte
for (i in 1:10){
  fold<- paste("fold", sprintf("%02d", i), sep = "_")
  assign(fold,data.frame(cbind(site_id=as.numeric(names(gbm.tune$fold.models[[i]]$data$y)),
    y=gbm.tune$fold.models[[i]]$data$y,fitted=plogis(gbm.tune$fold.models[[i]]$fit))))
}
neophyte3.obsfitted.folds = list(fold_01,fold_02,fold_03,fold_04,fold_05,fold_06,fold_07,fold_08,fold_09,fold_10)





#### MEMs ####


# MEM1 species richness
s=4
df=data.frame(GbmPlantsPreds[,c(s+3,9:ncol(GbmPlantsPreds))]) # without site_id
tuneParameters=gbm.tune.result.tree1000[gbm.tune.result.tree1000$y==s,]
tuneParameter=tuneParameters[1,] # test individual settings!

rm(.Random.seed, envir=globalenv())
set.seed(2020)     # reproducibility
    # train model
  gbm.tune <- gbm.step(
    data=df, gbm.x = 2:ncol(df), gbm.y = 1,
    family = "poisson",
    learning.rate = tuneParameter$learning.rate,
    bag.fraction = tuneParameter$bag.fraction,
    tree.complexity = tuneParameter$tree.complexity,
    n.folds=10,
    verbose=F
)
# ok!
explainedDeviance=(gbm.tune$self.statistics$mean.null-gbm.tune$cv.statistics$deviance.mean)/gbm.tune$self.statistics$mean.null
gbm.tune.result.1[s,] = c(tuneParameter[1:6],explainedDeviance,tuneParameter[7:12])
gbm.fits[[s]] = gbm.tune 

# MEM2 species richness
s=5
df=data.frame(GbmPlantsPreds[,c(s+3,9:ncol(GbmPlantsPreds))]) # without site_id
tuneParameters=gbm.tune.result.tree1000[gbm.tune.result.tree1000$y==s,]
tuneParameter=tuneParameters[1,] # test individual settings!

rm(.Random.seed, envir=globalenv())
set.seed(2020)     # reproducibility
    # train model
  gbm.tune <- gbm.step(
    data=df, gbm.x = 2:ncol(df), gbm.y = 1,
    family = "poisson",
    learning.rate = tuneParameter$learning.rate,
    bag.fraction = tuneParameter$bag.fraction,
    tree.complexity = tuneParameter$tree.complexity,
    n.folds=10,
    verbose=F
)
# ok!
explainedDeviance=(gbm.tune$self.statistics$mean.null-gbm.tune$cv.statistics$deviance.mean)/gbm.tune$self.statistics$mean.null
gbm.tune.result.1[s,] = c(tuneParameter[1:6],explainedDeviance,tuneParameter[7:12])
gbm.fits[[s]] = gbm.tune 




gbm.fittedValues.selectedNeophytes = fittedValues
gbm.fittedValues.10folds.selectedNeophytes = list(neophyte1.obsfitted.folds,neophyte2.obsfitted.folds,neophyte3.obsfitted.folds)
```





# Predicted vs. observed
```{r}
df = GbmPlantsPreds
#S-SDM
fittedValues.SSDM = gbm.fits[[1]]$fitted+gbm.fits[[2]]$fitted+gbm.fits[[3]]$fitted
cor(df$NfreqNeophytes,fittedValues.SSDM)
#[1] 0.6552167

# MEM 1
fittedValues.MEM1 = gbm.fits[[4]]$fitted
cor(df$NfreqNeophytes,fittedValues.MEM1)
#[1] 0.6730885

# MEM 2
fittedValues.MEM2 = gbm.fits[[5]]$fitted
cor(df$NtotalNeophytes,fittedValues.MEM2)
#[1] 0.7424807

##### Plot
my.formula=y~x

x = c(fittedValues.SSDM,fittedValues.MEM1,fittedValues.MEM2)
y = c(df$NfreqNeophytes,df$NfreqNeophytes,df$NtotalNeophytes)
modeltype=c(rep("S-SDM",143),rep("MEM1",143),rep("MEM2",143))

dat= data.frame(x,y,modeltype)
dat$modeltype = factor(dat$modeltype,levels = c("S-SDM", "MEM1", "MEM2"))

##################
dat1=dat%>%dplyr::filter(modeltype=="S-SDM")
summary(lm(y~x,data=dat1))
# Residuals:
#      Min       1Q   Median       3Q      Max 
# -1.36491 -0.42042 -0.09585  0.45233  1.63509 
# 
# Coefficients:
#             Estimate Std. Error t value Pr(>|t|)    
# (Intercept)  -0.5248     0.1380  -3.802 0.000213 ***
# x             1.6571     0.1609  10.299  < 2e-16 ***
# ---
# Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
# 
# Residual standard error: 0.6073 on 141 degrees of freedom
# Multiple R-squared:  0.4293,	Adjusted R-squared:  0.4253 
# F-statistic: 106.1 on 1 and 141 DF,  p-value: < 2.2e-16

dat2=dat%>%dplyr::filter(modeltype=="MEM1")
summary(lm(y~x,data=dat2))
# Residuals:
#     Min      1Q  Median      3Q     Max 
# -1.3009 -0.4293 -0.0736  0.4298  1.6991 
# 
# Coefficients:
#             Estimate Std. Error t value Pr(>|t|)    
# (Intercept)  -0.1952     0.1044  -1.869   0.0637 .  
# x             1.3357     0.1236  10.807   <2e-16 ***
# ---
# Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
# 
# Residual standard error: 0.5945 on 141 degrees of freedom
# Multiple R-squared:  0.453,	Adjusted R-squared:  0.4492 
# F-statistic: 116.8 on 1 and 141 DF,  p-value: < 2.2e-16

dat3=dat%>%dplyr::filter(modeltype=="MEM2")
summary(lm(y~x,data=dat3))

# Residuals:
#     Min      1Q  Median      3Q     Max 
# -1.5947 -0.5144 -0.1623  0.4856  4.4379 
# 
# Coefficients:
#             Estimate Std. Error t value Pr(>|t|)    
# (Intercept) -0.14553    0.12377  -1.176    0.242    
# x            1.18228    0.08983  13.162   <2e-16 ***
# ---
# Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
# 
# Residual standard error: 0.9409 on 141 degrees of freedom
# Multiple R-squared:  0.5513,	Adjusted R-squared:  0.5481 
# F-statistic: 173.2 on 1 and 141 DF,  p-value: < 2.2e-16


dat_lm <-dat  %>%
  group_by(modeltype) %>%
  nest() %>%
  # Fit linear model
  mutate(Mod = map(data, ~lm(y ~ x, data = .x))) %>%
  # Get the R2
  mutate(R2 = map_dbl(Mod, ~round(summary(.x)$r.squared, 2)),
         p = map_dbl(Mod, ~round(summary(.x)$coefficients[2,4], 3))) 

ggplot(data = dat, aes(x = x, y = y)) +
   geom_smooth(method = "lm", se=FALSE, color="black", formula = my.formula) +
   geom_point() + 
  facet_wrap(.~modeltype,scales="free")+
  labs(y = "observed", x = "fitted") +
  # Add label
  geom_richtext(data = dat_lm, 
             aes(x = Inf, y = Inf, 
                 label = paste0("r<sup>2</sup>= ", R2, ", p<0.001",sep = " ")),
             vjust = 1.8, hjust = 2,inherit.aes = FALSE,fill = NA, label.color = NA) +
theme_bw() + mytheme
```




















# Final table - Individual GBM
```{r}
explainedDeviance = gbm.tune.result.1$explainedDeviance
cv.correlation.mean = gbm.tune.result.1$cv.correlation.mean
AUCtrain.mean = gbm.tune.result.1$AUCtrain.mean


GBMindv.explainedDeviance = round(mean(explainedDeviance[1:3]),2)
#GBMindv.cv.deviance.sd  = paste("(",round(min(cv.deviance.mean[1:3]),2),")",sep="")
GBMindv.explainedDeviance.range  = paste(round(range(explainedDeviance[1:3])[1],2),round(range(explainedDeviance[1:3])[2],2),sep="-")

GBMindv.cv.correlation.mean = round(mean(cv.correlation.mean[1:3]),2)
#GBMindv.cv.correlation.sd  = paste("(",round(sd(cv.correlation.mean[1:3]),2),")",sep="")
GBMindv.cv.correlation.range  = paste(round(range(cv.correlation.mean[1:3])[1],2),round(range(cv.correlation.mean[1:3])[2],2),sep="-")

GBMindv.AUCtrain.mean =  round(mean(AUCtrain.mean[1:3]),2)
#GBMindv.AUCtrain.sd  = paste("(",round(sd(AUCtrain.mean[1:3]),2),")",sep="")
GBMindv.AUCtrain.range  = paste(round(range(AUCtrain.mean[1:3])[1],2),round(range(AUCtrain.mean[1:3])[2],2),sep="-")

GBMdeviance = cbind(mean=GBMindv.explainedDeviance,range=GBMindv.explainedDeviance.range,
      MEM1=round(explainedDeviance[4],2),MEM2=round(explainedDeviance[5],2))
GBMcorr = cbind(mean=GBMindv.cv.correlation.mean,range=GBMindv.cv.correlation.range,
      MEM1=round(cv.correlation.mean[4],2),MEM2=round(cv.correlation.mean[5],2))
GBMauc = cbind(mean=GBMindv.AUCtrain.mean,range=GBMindv.AUCtrain.range,"","")


GBMresult=data.frame(rbind(GBMdeviance,GBMcorr,GBMauc))
GBMresult
```




# S-SDM - Observed and predicted from GBM
Performance measures are applied on averaged fitted values from the 10 folds. 

```{r}
fittedValues = gbm.fittedValues.10folds.selectedNeophytes # results from gbm models

ss.fit=data.frame(matrix(0,nrow=143,ncol=7))
colnames(ss.fit)=c("site_id","sp1.obs","sp1.preval","sp2.obs","sp2.preval","sp3.obs","sp3.preval")

# summarise fitted values from 10-folds from same "sites"
sf.fit = rep(NA,6)
Nsf.fit = rep(1:4,each=2)
for (o in 1:143){
for (s in 1:3){
dat=dplyr::bind_rows(lapply(1:10,function(f) data.frame(fittedValues[[s]][[f]][fittedValues[[s]][[f]]$site_id==o,])))
sf.fit[which(Nsf.fit==s)]=c(dat$y[1],mean(dat$fitted,na.rm=T))}
ss.fit[o,]=c(o,rbind(sf.fit))}

ss.fit=na.omit(ss.fit)
gbm.ss.fit = ss.fit
```

C-index
```{r}
ss.fit = gbm.ss.fit
Cindex = rep(0,nrow(ss.fit))
for (s in 1:nrow(ss.fit)){
# get obs and fitted from folds
# observation used in fold for all three species
obs.SR=c(ss.fit$sp1.obs[s],ss.fit$sp2.obs[s],ss.fit$sp3.obs[s])

# Probability for each species to occur at the site(s)/prevalence
sp.preval=c(ss.fit$sp1.preval[s],ss.fit$sp2.preval[s],ss.fit$sp3.preval[s])
 


# print(all.equal((fitted(m)), predict(m,type="response")))
PresSpecies=which(obs.SR==1)
AbsSpecies=which(obs.SR==0)


obsPres =length(which(obs.SR==1))
obsAbs =length(which(obs.SR==0))
Pairs=obsPres*obsAbs
if (Pairs>0){
  Combinations=data.frame(crossing(PresSpecies,AbsSpecies))
  Comparisons = rep(0,nrow(Combinations))

  for(i in 1:nrow(Combinations)){if(sp.preval[Combinations[i,1]]>sp.preval[Combinations[i,2]]){Comparisons[i]=1}}
  Cindex[s] = sum(Comparisons)/Pairs
  }
if (Pairs==0){Cindex[s] = NA}

}
  
  
mean(as.vector(na.omit(Cindex)))
sd(as.vector(na.omit(Cindex)))
range(as.vector(na.omit(Cindex)))
gbm.Cindex = Cindex
```



```{r}
ss.fit = gbm.ss.fit

sp.pool = 3 # species to occur at a site (Regional species pool)


simSR.selectedNeophytesSite = as_tibble(data.frame(matrix(0,ncol=5,nrow=nrow(ss.fit))))
colnames(simSR.selectedNeophytesSite)=c("species","site_id","obsSR","realSR","expectedSR")

loopN = rep(1:nrow(ss.fit),each=sp.pool)

# Tests:
#sp.pool <- round(runif(1, min=50, max=100)) #Random number of potental species to occur at a site (Regional species pool
#sp.preval <- round(runif(sp.pool, min=0, max=1), 3) #Random probability for each species to occur at the site(s)/prevalence

rep <- 1000 #Number of times the binomial distribution is drawn to create the "observed" species richness
run=1
for (s in 1:nrow(ss.fit)){
# get obs and fitted from folds
# observation used in fold for all four species
obs.SR=c(ss.fit$sp1.obs[s],ss.fit$sp2.obs[s],ss.fit$sp3.obs[s])

# Probability for each species to occur at the site(s)/prevalence
sp.preval=c(ss.fit$sp1.preval[s],ss.fit$sp2.preval[s],ss.fit$sp3.preval[s])

#Site parameters calculation based on the input data
#expected.SR.sum <- sum(sp.preval) #The expected species richness
set.seed(2020)
expected.SR <- poibin::dpoibin(1:sp.pool, sp.preval) #The expected probability for all possible SR from 1 to sp.pool based on sp.preval

#Simple model to create n=rep, realisations of the probability distribution
obs.SR.binom <- NULL
for(i in 1:rep){
  set.seed(2020)
  obs.SR.binom <- c(obs.SR.binom, sum(rbinom(sp.pool,1,sp.preval)))
}
#Histogram of the observed SR based on independent Bernoulli trials
SR.hist <- hist(obs.SR.binom, breaks=0:sp.pool) # max. sp.pool species can occur in hist()
SR.hist$counts <- SR.hist$counts/rep #Standardisation with the number of repetition
real.obs.SR = SR.hist$counts

simSR.selectedNeophytesSite[which(loopN==run),] = cbind(1:sp.pool,rep(s,sp.pool),obs.SR,real.obs.SR,expected.SR)
run=run+1

}



# WITH REAL OBS (sim with rbinom): 10 folds representiv for site = 1 (obs 1)
corSR.selectedNeophytesSite = simSR.selectedNeophytesSite %>% dplyr::select(site_id,realSR,expectedSR) %>% dplyr::group_by(site_id) %>% dplyr::summarise(corSR=cor(realSR,expectedSR))
# final results for 128 sites - based on 143 with 10 folds cv
mean(corSR.selectedNeophytesSite$corSR)
range(corSR.selectedNeophytesSite$corSR)


# WITH observed OBS: 10 folds representiv for site = 1 (obs 1)
corSR.selectedNeophytesSite = simSR.selectedNeophytesSite %>% dplyr::select(site_id,obsSR,expectedSR) %>% dplyr::group_by(site_id) %>% dplyr::summarise(corSR=ifelse(var(obsSR)==0,NA,abs(cor(obsSR,expectedSR))))

# final results for 128 sites - based on 143 with 10 folds cv
mean(corSR.selectedNeophytesSite$corSR,na.rm=T)
range(corSR.selectedNeophytesSite$corSR,na.rm=T)


gbm.simSR.selectedNeophytesSite = simSR.selectedNeophytesSite # save

```

```{r}
save(GBMresult,gbm.tune.result.1,gbm.fits,
gbm.fittedValues.selectedNeophytes,
gbm.simSR.selectedNeophytesSite,gbm.Cindex,gbm.ss.fit,
file="results/results.gbm.RData")
```


# plots
```{r}
gbm.plot.fits(final.gbm.MEM)



df2 <- summary(final.gbm.MEM) %>% 
  dplyr::rename("variable" = var) %>% 
  dplyr::arrange(rel.inf) %>%
  dplyr::mutate(variable = forcats::fct_inorder(variable))

ggplot2::ggplot(df2) +
  geom_col(aes(x = variable, y = rel.inf),
           col = "black", show.legend = F) +
  coord_flip() +
  scale_fill_grey() +
  theme_bw()

# library(vip)
# devtools::install_github("koalaverse/vip")
vip::vip(gbm.fit.final)
```
