==> r-code/05_3_statistical_analyses_BRMS.Rmd <==
---
title: "workingdata"
output: pdf_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
source("r-code/00_functions.R")
load("data/workingdata.RData")
using("ggplot2","tidyverse","knitr","brms","sjPlot","gridExtra","future","ggpmisc","ggpubr","ggtext","loo","bayesplot")
plan(multicore)

mytheme <-  theme(axis.title = element_text(size =18),
        axis.text = element_text(size =14),
        strip.text= element_text(size = 18),
        legend.position = "none",
        plot.title = element_text(hjust = 0.5))
```



https://www.monicaalexander.com/posts/2020-28-02-bayes_viz/


```{r}
Preds = preds %>% dplyr::select(site_id,plot_id,ring1,slope,herbAbund.4,humus.0)
GbmPlantsPreds = as_tibble(dplyr::left_join(plants,Preds,by=c("site_id","plot_id")))
```




```{r}
df = GbmPlantsPreds[,-c(1,2,3)]
fit=list()

for (s in 1:5){
dat=data.frame(cbind(y=df[,s],df[,6:ncol(df)],GbmPlantsPreds[,1]))
colnames(dat)[1]="y"

### family!
# model are run separatly, for single species logit, for community negbinomial
#family=bernoulli(link = "logit")

#if(s %in% c(4,5)) {family = "poisson"}
if(s %in% c(4,5)) {family = "negbinomial"}

set.seed(2020)
fit[[s]] <- brm(y ~ ring1+slope+herbAbund.4+humus.0+(1|site_id),data=dat,family=family,iter = 30000,chains=4,control = list(adapt_delta = 0.99,max_treedepth=15))
}


# show summary
lapply(fit, function(x)summary(x))


# 10 fold cv
for (i in 1:5){
set.seed(2020)
  # use the future package for parallelization

  fit.kfold=brms::kfold(fit[[i]], k=10,save_fits = TRUE)
assign(paste("fit.kfold",i,"",sep=""),fit.kfold)
}

#fit.kfold = NULL
fit.kfold.iter30000 = list(fit.kfold1,fit.kfold2,fit.kfold3,fit.kfold4,fit.kfold5)


fittedValues.m1 = fitted(fit[[1]])
fittedValues.m2 = fitted(fit[[2]])
fittedValues.m3 = fitted(fit[[3]])
fittedValues.m4 = fitted(fit[[4]])
fittedValues.m5 = fitted(fit[[5]])
```


# plot individual models - sjplot
```{r}
mytheme <-  theme(axis.title = element_text(size =20),
        axis.text = element_text(size =20),
        strip.text= element_text(size = 18),
        legend.position = "none",
        plot.title = element_text(hjust = 0.5,size=20,face = "italic"))

mylabels = c("low levels of organic matter\n(absent or rare)",">50% coverage of herbs in plot\n(herb abundance = abundant)","slope","Zone 1","Intercept")

pp1 <- plot_model(fit[[1]],transform = NULL, title = "Cynodon dactylon",vline.color = "gray90",show.intercept = T,colors=c("gray70","gray5"))+theme_bw()+mytheme
levels(pp1$data$term)=mylabels
  
pp2 <- plot_model(fit[[2]],transform = NULL, title = "Megathyrsus maximus",vline.color = "gray90",show.intercept = T,colors=c("gray70","gray5"))+theme_bw()+mytheme+theme(axis.text.y = element_blank())
levels(pp2$data$term)=mylabels

pp3 <- plot_model(fit[[3]],transform = NULL, title = "Ricinus communis",vline.color = "gray90",show.intercept = T,colors=c("gray70","gray5"))+theme_bw()+mytheme+theme(axis.text.y = element_blank())
levels(pp3$data$term)=mylabels

ggpubr::ggarrange(pp1, pp2,pp3, widths = c(3,1.5,1.5),nrow=1)
```

# plot individual models - bayesplot
```{r}
mytheme1 <-  theme(axis.title = element_text(size =20),
        axis.text = element_text(size =20),
        strip.text= element_text(size = 18),
        legend.position = "none",
        plot.title = element_text(hjust = 0.5,size=20,face = "italic"))


mytheme2 <-  theme(axis.title = element_text(size =20),
        axis.ticks.y=element_blank(),
        axis.text.y=element_blank(),
        axis.text = element_text(size =20),
        strip.text= element_text(size = 18),
        legend.position = "none",
        plot.title = element_text(hjust = 0.5,size=20,face = "italic"))

pars= c("b_Intercept","b_ring11","b_slope","b_herbAbund.41","b_humus.01")

Div = function(x) {x/10}

pp1 <- mcmc_intervals(fit[[1]],pars=pars,transformations="Div")+theme_bw()+mytheme1 + labs(title = "Cynodon dactylon") +xlim(-5,5)  #scale_y_discrete(labels= mylabels) 

pp2 <- mcmc_intervals(fit[[2]],pars=pars)+theme_bw()+mytheme2 + labs(title = "Megathyrsus maximus")+xlim(-5,5) 

pp3 <- mcmc_intervals(fit[[3]],pars=pars)+theme_bw()+mytheme2 + labs(title = "Ricinus communis")+xlim(-5,5)

ggpubr::ggarrange(pp1, pp2,pp3, widths = c(2.3,1.2,1.2),nrow=1)
```




####
```{r}
#summary(fit[[s]])

dharmaPlots=list()
for (s in 1:5){
set.seed(2021)
preds <- posterior_predict(fit[[s]], nsamples = 250, summary = FALSE)
preds <- t(preds)


res <- createDHARMa(simulatedResponse = preds, 
                    fittedPredictedResponse = apply(preds, 1, mean), 
                    observedResponse = fit[[s]]$data$y)

dharmaPlots[[s]]=plot(res)
}

```



# Fitted vs. observed
```{r}
df = GbmPlantsPreds
#S-SDM
fittedValues.SSDM = fittedValues.m1[,'Estimate']+fittedValues.m2[,'Estimate']+fittedValues.m3[,'Estimate']
plot(df$NfreqNeophytes,fittedValues.SSDM)
cor(df$NfreqNeophytes,fittedValues.SSDM)
# [1] 0.7286541

# MEM 1
fittedValues.MEM1 = fittedValues.m4[,'Estimate']
plot(df$NfreqNeophytes,fittedValues.MEM1)
cor(df$NfreqNeophytes,fittedValues.MEM1)
#[1] 0.6520213

# MEM 2
fittedValues.MEM2 = fittedValues.m5[,'Estimate']
cor(df$NtotalNeophytes,fittedValues.MEM2)
#[1] 0.8196173

##### Plot
my.formula=y~x

x = c(fittedValues.SSDM,fittedValues.MEM1,fittedValues.MEM2)
y = c(df$NfreqNeophytes,df$NfreqNeophytes,df$NtotalNeophytes)
modeltype=c(rep("S-SDM",143),rep("MEM1",143),rep("MEM2",143))

dat= data.frame(x,y,modeltype)
dat$modeltype = factor(dat$modeltype,levels = c("S-SDM", "MEM1", "MEM2"))

##################
dat1=dat%>%dplyr::filter(modeltype=="S-SDM")
summary(lm(y~x,data=dat1))
# Residuals:
#     Min      1Q  Median      3Q     Max 
# -1.6422 -0.3326 -0.1271  0.3850  1.6238 
# 
# Coefficients:
#             Estimate Std. Error t value Pr(>|t|)    
# (Intercept) -0.07030    0.08267   -0.85    0.397    
# x            1.08230    0.08567   12.63   <2e-16 ***
# ---
# Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
# 
# Residual standard error: 0.5506 on 141 degrees of freedom
# Multiple R-squared:  0.5309,	Adjusted R-squared:  0.5276 
# F-statistic: 159.6 on 1 and 141 DF,  p-value: < 2.2e-16

dat2=dat%>%dplyr::filter(modeltype=="MEM1")
summary(lm(y~x,data=dat2))
# Residuals:
#     Min      1Q  Median      3Q     Max 
# -1.3700 -0.4194 -0.2097  0.4439  1.9386 
# 
# Coefficients:
#             Estimate Std. Error t value Pr(>|t|)    
# (Intercept) -0.01389    0.09438  -0.147    0.883    
# x            1.01699    0.09959  10.211   <2e-16 ***
# ---
# Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
# 
# Residual standard error: 0.6095 on 141 degrees of freedom
# Multiple R-squared:  0.4251,	Adjusted R-squared:  0.4211 
# F-statistic: 104.3 on 1 and 141 DF,  p-value: < 2.2e-16

dat3=dat%>%dplyr::filter(modeltype=="MEM2")
summary(lm(y~x,data=dat3))
# Residuals:
#     Min      1Q  Median      3Q     Max 
# -2.1061 -0.4052 -0.1780  0.4926  4.3181 
# 
# Coefficients:
#             Estimate Std. Error t value Pr(>|t|)    
# (Intercept) -0.07764    0.09712  -0.799    0.425    
# x            1.06971    0.06297  16.988   <2e-16 ***
# ---
# Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
# 
# Residual standard error: 0.8048 on 141 degrees of freedom
# Multiple R-squared:  0.6718,	Adjusted R-squared:  0.6694 
# F-statistic: 288.6 on 1 and 141 DF,  p-value: < 2.2e-16


dat_lm <-dat  %>%
  group_by(modeltype) %>%
  nest() %>%
  # Fit linear model
  mutate(Mod = map(data, ~lm(y ~ x, data = .x))) %>%
  # Get the R2
  mutate(R2 = map_dbl(Mod, ~round(summary(.x)$r.squared, 2)),
         p = map_dbl(Mod, ~round(summary(.x)$coefficients[2,4], 3))) 

ggplot(data = dat, aes(x = x, y = y)) +
   geom_smooth(method = "lm", se=FALSE, color="black", formula = my.formula) +
   geom_point() + 
  facet_wrap(.~modeltype,scales="free")+
  labs(y = "observed", x = "fitted") +
  # Add label
  geom_richtext(data = dat_lm, 
             aes(x = Inf, y = Inf, 
                 label = paste0("r<sup>2</sup>= ", R2, ", p<0.001",sep = " ")),
             vjust = 1.8, hjust = 2,inherit.aes = FALSE,fill = NA, label.color = NA) +
theme_bw() + mytheme
```



# Get 10cv performance criteria: Looic, Bayes R2
```{r}
#fit.kfold=fit.kfold.iter30000
brms.perfCritera = data.frame(matrix(0,ncol=13,nrow=5))
colnames(brms.perfCritera) = c("model",
                               "looic.mean","looic.min","looic.max",
                               "bayes_R2.mean","bayes_R2.min","bayes_R2.max",
                               "AUC.mean","AUC.min","AUC.max",
                               "Cor.mean","Cor.min","Cor.max")

brmLooicFold = rep(NA,10) # looic
brmR2Fold    = rep(NA,10) # R2
brmAUCFold   = rep(NA,10) # AUC
brmCor       = rep(NA,10) # Cor fitted train vs. test

nFolds=10

saveAllWarnings(
for (i in 5){
kfold.m = fit.kfold[[i]]

for(f in 1:nFolds){
y = kfold.m$fits[[f]]$data$y
fittedValues = predict(kfold.m$fits[[f]],type = "response")[ , "Estimate"]   

### AUC ### for binomial single SDM
    if(i<4){
     
set.seed(2020)
brmAUCFold[f]=as.numeric(pROC::roc(response = y, predictor =  fittedValues, plot = FALSE, print.auc = TRUE)$auc)
    }


   # looic & R2
   kfold.m$fits[[f]] <- brms::add_criterion(kfold.m$fits[[f]], c("loo", "bayes_R2","loo_R2"))   
   brmLooicFold[f] = kfold.m$fits[[f]]$criteria$loo$estimates[3,1] # looic
   brmR2Fold[f]    = mean(kfold.m$fits[[f]]$criteria$bayes_R2,na.rm=T) # R2
   brmCor[f]       = cor(y,fittedValues)
  }

brms.perfCritera$model[i]=i
brms.perfCritera$looic.mean[i]= mean(brmLooicFold,na.rm=T)
brms.perfCritera$looic.min[i] = min(brmLooicFold,na.rm=T)
brms.perfCritera$looic.max[i] = max(brmLooicFold,na.rm=T)

brms.perfCritera$bayes_R2.mean[i] = mean(brmR2Fold,na.rm=T)
brms.perfCritera$bayes_R2.min[i]  = min(brmR2Fold,na.rm=T)
brms.perfCritera$bayes_R2.max[i]  = max(brmR2Fold,na.rm=T)

brms.perfCritera$Cor.mean[i] = mean(brmCor,na.rm=T)
brms.perfCritera$Cor.min[i]  = min(brmCor,na.rm=T)
brms.perfCritera$Cor.max[i]  = max(brmCor,na.rm=T)

if(i<4){
  brms.perfCritera$AUC.mean[i]= mean(brmAUCFold,na.rm=T)
  brms.perfCritera$AUC.max[i] = max(brmAUCFold,na.rm=T)
  brms.perfCritera$AUC.min[i] = min(brmAUCFold,na.rm=T)
}}
, logFile=paste("results/2021.01.24-BRMS.neophyte",i,".K10CV.warninglog.R",sep=""))

brms.perfCritera

mean(brms.perfCritera$bayes_R2.mean[1:3]);brms.perfCritera$bayes_R2.mean[1:3] # BayesR
mean(brms.perfCritera$Cor.mean[1:3]);brms.perfCritera$Cor.mean[1:3] # Cor
mean(brms.perfCritera$AUC.mean[1:3]);brms.perfCritera$AUC.mean[1:3] # AUC
```



# S-SDM
## extract predictions from individual folds for 4 selected neophytes
```{r}
obsfitted.folds= list()
nFolds=10
for (i in 1:3){ # 4 single SDMs
kfold.m = fit.kfold[[i]]

 for(f in 1:nFolds){
      y = kfold.m$fits[[f]]$data$y
      set.seed(2020)
      p = predict(kfold.m$fits[[f]],type = "response")[,"Estimate"]
      site_id=kfold.m$fits[[f]]$data$site_id
      obsfitted.folds[[f]]=data.frame(site_id,y,fitted=p)
 }

assign(paste("neophyte",i,".obsfitted.folds",sep=""),obsfitted.folds)
}

brms.fittedValues.10folds.selectedNeophytes = list(neophyte1.obsfitted.folds,neophyte2.obsfitted.folds,neophyte3.obsfitted.folds)
```

```{r}
fittedValues = brms.fittedValues.10folds.selectedNeophytes # results from gbm models

ss.fit=data.frame(matrix(0,nrow=143,ncol=7))
colnames(ss.fit)=c("site_id","sp1.obs","sp1.preval","sp2.obs","sp2.preval","sp3.obs","sp3.preval")

sf.fit = rep(NA,6)
Nsf.fit = rep(1:3,each=2)
for (o in 1:143){
for (s in 1:3){
dat=dplyr::bind_rows(lapply(1:10,function(f) data.frame(fittedValues[[s]][[f]][fittedValues[[s]][[f]]$site_id==o,])))
sf.fit[which(Nsf.fit==s)]=c(dat$y[1],mean(dat$fitted,na.rm=T))}
ss.fit[o,]=c(o,rbind(sf.fit))}

ss.fit=na.omit(ss.fit)
brms.ss.fit = ss.fit
```


```{r}
ss.fit = brms.ss.fit
Cindex = rep(0,nrow(ss.fit))
for (s in 1:nrow(ss.fit)){
# get obs and fitted from folds
# observation used in fold for all four species
obs.SR=c(ss.fit$sp1.obs[s],ss.fit$sp2.obs[s],ss.fit$sp3.obs[s],ss.fit$sp4.obs[s])

# Probability for each species to occur at the site(s)/prevalence
sp.preval=c(ss.fit$sp1.preval[s],ss.fit$sp2.preval[s],ss.fit$sp3.preval[s],ss.fit$sp4.preval[s])
 
Cindex[s]=cAUC.fct(p=sp.preval,y=obs.SR)
}
  
  
mean(as.vector(na.omit(Cindex)))
```

```{r}
save(fit,fit.kfold.iter30000,brms.perfCritera,brms.fittedValues.10folds.selectedNeophytes,
     Cindex,dharmaPlots,
     file="results/2021.01.24-result.brms.iter30000.RData")
```
